chore: move tests covering resource pool to the right file (#9447)
